Move abbreviation parsing into separate module
